requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
h11==0.14.0
idna==3.7
migrate==0.3.7
oauthlib==3.2.2
outcome==1.3.0.post0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
python3-openid==3.2.0
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.21.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.1
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.1
websockets==12.0
wsproto==1.2.0
